Move media player to separate module
Should enable us to run tools without installing qt.
