+++ src/usdb_syncer/media_player.py
@@ -1,0 +1,14 @@
+from PySide6 import QtMultimedia
+
+
+class _MediaPlayer:
+    player: QtMultimedia.QMediaPlayer | None = None
+
+
+def media_player() -> QtMultimedia.QMediaPlayer:
+    if _MediaPlayer.player is None:
+        _MediaPlayer.player = QtMultimedia.QMediaPlayer()
+        _MediaPlayer.player.setAudioOutput(
+            QtMultimedia.QAudioOutput(_MediaPlayer.player)
+        )
+    return _MediaPlayer.player
